refactor(CodeEditor): consolidate module loading logic with useModules hook
- Replaced the previous module loading approach with a new `useModules` hook to streamline the management of core and formatting modules.
- Removed redundant code related to missing module checks and lazy loading, enhancing clarity and maintainability.
- Updated the CodeEditor component to utilize the new hook, ensuring consistent module handling across functionalities.
- Improved TypeScript definitions for better type safety and clarity in module management.
